+++ html-report/src/components/common/ExecutionIcon.vue
