+++ html-report/src/components/details/Block.vue
