+++ html-report/src/components/details/CodeBlock.vue
