+++ html-report/src/components/details/DetailsHeader.vue
