+++ html-report/src/components/details/DetailsSections.vue
