+++ html-report/src/components/details/DurationLabel.vue
