+++ html-report/src/components/details/Section.vue
