+++ html-report/src/components/details/keys.ts
