+++ html-report/src/components/header/DarkModeSwitch.vue
