+++ html-report/src/components/sidebar/ToolBarIcon.vue
